put & delete stuff

+++ src/main/java/hu/bme/aut/javaweb/forum/service/UserService.java
@@ -4,6 +4,7 @@ import hu.bme.aut.javaweb.forum.datasource.UserDataSource;
 import hu.bme.aut.javaweb.forum.model.User;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
@@ -15,7 +16,15 @@ public class UserService {
         this.userDataSource = userDataSource;
     }
 
-    public User getUserById(Long id) {
+    public List<User> getAllUsers() {
+        return userDataSource.findAll();
+    }
+
+    public User getUserById(Long id, Long userId, Boolean admin) {
+        if (id != userId && !admin) {
+            throw new IllegalArgumentException("wrong userId");
+        }
+
         Optional<User> user = userDataSource.findById(id);
 
         if (user.isEmpty()) {
@@ -25,15 +34,19 @@ public class UserService {
         return user.get();
     }
 
-    public User createUser(User user) {
-        return userDataSource.save(user);
-    }
-
     public User updateUser(User user) {
         return userDataSource.save(user);
     }
 
-    public void deleteUserById(Long id) {
+    public void deleteUserById(Long id, Long userId, Boolean admin) {
+        if (id != userId && !admin) {
+            throw new IllegalArgumentException("wrong userId");
+        }
+
+        if (id == userId && admin) {
+            throw new IllegalArgumentException("can not delete admin");
+        }
+
         userDataSource.deleteById(id);
     }
 }
